Move helper only used in tests to test folder
