+++ test/cash_register_test.exs
@@ -3,8 +3,6 @@ defmodule CashRegisterTest do
 
   import CashRegister.DenominationHelpers
 
-  alias CashRegister.Parser
-
   describe "process_transaction/3" do
     test "processes single transaction" do
       assert {:ok, "3 quarters,1 dime,3 pennies"} = CashRegister.process_transaction(212, 300)
@@ -22,7 +20,7 @@ defmodule CashRegisterTest do
     test "processes transaction with divisible-by-3 change" do
       assert {:ok, result} = CashRegister.process_transaction(101, 200)
 
-      assert {:ok, denominations} = Parser.parse_change_result(result)
+      assert {:ok, denominations} = parse_change_result(result)
 
       assert_correct_total(denominations, 99)
     end
@@ -31,7 +29,7 @@ defmodule CashRegisterTest do
       # 90 is divisible by 5, so with divisor: 5 it uses Randomized
       assert {:ok, result} = CashRegister.process_transaction(10, 100, divisor: 5)
 
-      assert {:ok, denominations} = Parser.parse_change_result(result)
+      assert {:ok, denominations} = parse_change_result(result)
 
       assert_correct_total(denominations, 90)
     end
@@ -98,7 +96,7 @@ defmodule CashRegisterTest do
       assert [result] = results
 
       # Should still total to 90 cents
-      assert {:ok, denominations} = Parser.parse_change_result(result)
+      assert {:ok, denominations} = parse_change_result(result)
       assert_correct_total(denominations, 90)
     end
   end
